+++ packages/iov-bns/src/encode.spec.ts
@@ -65,18 +65,27 @@ import { appendSignBytes } from "./util";
 const { fromHex } = Encoding;
 
 describe("Encode", () => {
-  it("encode pubkey", () => {
-    const pubkey = encodePubkey(pubJson);
-    const encoded = codecImpl.crypto.PublicKey.encode(pubkey).finish();
-    // force result into Uint8Array for tests so it passes
-    // if buffer of correct type as well
-    expect(Uint8Array.from(encoded)).toEqual(pubBin);
+  describe("encodePubkey", () => {
+    it("can encode a pubkey", () => {
+      const pubkey = encodePubkey(pubJson);
+      const encoded = codecImpl.crypto.PublicKey.encode(pubkey).finish();
+      // force result into Uint8Array for tests so it passes
+      // if buffer of correct type as well
+      expect(Uint8Array.from(encoded)).toEqual(pubBin);
+    });
+
+    it("throws for invalid size", () => {
+      const pubkey = { algo: Algorithm.Ed25519, data: fromHex("ab") as PubkeyBytes };
+      expect(() => encodePubkey(pubkey)).toThrowError(/invalid pubkey size/i);
+    });
   });
 
-  it("encode private key", () => {
-    const privkey = encodePrivkey(privJson);
-    const encoded = codecImpl.crypto.PublicKey.encode(privkey).finish();
-    expect(Uint8Array.from(encoded)).toEqual(privBin);
+  describe("encodePrivkey", () => {
+    it("can encode a privkey", () => {
+      const privkey = encodePrivkey(privJson);
+      const encoded = codecImpl.crypto.PrivateKey.encode(privkey).finish();
+      expect(Uint8Array.from(encoded)).toEqual(privBin);
+    });
   });
 
   describe("encodeAmount", () => {
@@ -160,13 +169,15 @@ describe("Encode", () => {
       nonce: 123 as Nonce,
       pubkey: {
         algo: Algorithm.Ed25519,
-        data: fromHex("00aa1122bbddffeeddcc") as PubkeyBytes,
+        data: fromHex("00aa1122bbddffeeddcc00aa1122bbddffeeddcc00aa1122bbddffeeddcc00aa") as PubkeyBytes,
       },
       signature: fromHex("aabbcc22334455") as SignatureBytes,
     };
     const encoded = encodeFullSignature(fullSignature);
     expect(encoded.sequence).toEqual(123);
-    expect(encoded.pubkey!.ed25519!).toEqual(fromHex("00aa1122bbddffeeddcc"));
+    expect(encoded.pubkey!.ed25519!).toEqual(
+      fromHex("00aa1122bbddffeeddcc00aa1122bbddffeeddcc00aa1122bbddffeeddcc00aa"),
+    );
     expect(encoded.signature!.ed25519).toEqual(fromHex("aabbcc22334455"));
   });